rename plugins utils to app_intergrations and remove old plugin utils
